chatbot/requirements.txt
streamlit
pymongo
requests
pandas
python-dotenv
ollama
feedparser
beautifulsoup4